+++ .travis.yml
@@ -13,11 +13,11 @@ python:
   - "3.3"
   - "3.4"
   - "3.5"
-  - "3.5-dev" # 3.5 development branch
+  #- "3.5-dev" # 3.5 development branch
   - "3.6"
-  - "3.6-dev" # 3.6 development branch
-  - "3.7-dev" # 3.7 development branch
-  - "nightly" # currently points to 3.7-dev
+  #- "3.6-dev" # 3.6 development branch
+  #- "3.7-dev" # 3.7 development branch
+  #- "nightly" # currently points to 3.7-dev
 
 # command to install dependencies
 #install:
