+++ README.md
@@ -1,8 +1,52 @@
 
-# UniMon
+UniMon
+======
 
 UniMon is a solution for unifying agent and agent-less servers together under a unified dashboard.
 
 * Code publicly hosted on GitHub: <https://github.com/pathaugen/unimon>
 * Utilizing Travis CI for builds: <https://travis-ci.org/pathaugen/unimon>
-* Running on Google App Engine (GAE): <xxx>
+* Running on Google App Engine (GAE): <https://pmd-unimon.appspot.com/>
+
+---------- ---------- ---------- ---------- ----------
+
+External Backend Libraries and Resources Included in Application
+----------------------------------------------------------------
+
+* GAE LIBRARY
+* AWS LIBRARY
+
+---------- ---------- ---------- ---------- ----------
+
+External Frontend Libraries and Resources Included in Application
+----------------------------------------------------------------
+
+* Font Awesome 4.7.0: <http://fontawesome.io/>
+* jQuery 3.1.1: <http://jquery.com/>
+* jQuery UI 1.12.1: <http://jqueryui.com/>
+* VISUALIZATION LIBRARY
+
+---------- ---------- ---------- ---------- ----------
+
+Feature List
+------------
+
+* Backend:
+  * Custom backend in Python hosted on Google App Engine
+  * Utilization of pure HTML/CSS/JS to keep future updates to app simple
+  * Login via Google credentials to backend dashboard
+  * Dashboard for viewing monitoring data
+  * Continuous Integration via Travis CI autodeploys on GitHub master branch checkin
+
+* Frontend:
+  * Responsive design
+  * xxx library for visualization of data
+
+---------- ---------- ---------- ---------- ----------
+
+Resources
+---------
+
+* Google App Engine: <https://cloud.google.com/appengine/>
+* Google Hosted Libraries: <https://developers.google.com/speed/libraries/>
+* Markdown Syntax: <http://daringfireball.net/projects/markdown/syntax>
