+++ unimon/constants.py
@@ -1,3 +1,4 @@
 
 EMAIL       = ''
 REDIRECT    = ''
+POSTVALUES  = ''
