+++ unimon/post_ajax.py
@@ -1,0 +1,17 @@
+
+#import json
+
+#from google.appengine.api import users
+
+import constants as c
+
+#from unimon.database import CheckInFH5
+
+# ---------- ---------- ---------- ---------- ---------- ---------- ---------- ---------- ---------- ---------- #
+def post_ajax(post_values):
+    c.POSTVALUES = post_values
+    ajax = ''
+    
+    return ajax
+# ---------- ---------- ---------- ---------- ---------- ---------- ---------- ---------- ---------- ---------- #
+
